Add support for nested group creation
Add support to find hosts in parent hostgroups as well. Eg: if a hosts belongs to a group in foreman like so 'prod/tier1/cache' it would be found either in 'foreman_hostgroup_prod', 'foreman_hostgroup_prod_tier1' and 'foreman_hostgroup_prod_tier1_cache' and not only in 'prod_tier1_cache'.